+++ nengo_spa/modules/tests/test_superposition.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 import nengo
-from nengo.utils.numpy import rmse
+from nengo.utils.numpy import rms
 
 import nengo_spa as spa
 from nengo_spa.vocabulary import VocabularyMap
@@ -43,8 +43,8 @@ def test_run(Simulator, algebra, seed):
     with Simulator(model) as sim:
         sim.run(0.2)
 
-    error = rmse(vocab.parse("(B+A).normalized()").v, sim.data[p][-1])
+    error = rms(vocab.parse("(B+A).normalized()").v - sim.data[p][-1])
     assert error < 0.1
 
-    error = rmse(vocab.parse("(A+A).normalized()").v, sim.data[p][100])
+    error = rms(vocab.parse("(A+A).normalized()").v - sim.data[p][100])
     assert error < 0.2